training-pipeline/vlm.py
# Research Paper Extraction with Vision-Language Models

import cv2
import numpy as np
from PIL import Image
import pytesseract
from transformers import ViltProcessor, ViltForQuestionAnswering

# Note: This notebook requires additional setup for the VILT model and pytesseract

# 1. PDF Parsing and Content Extraction

def extract_text_from_image(image_path: str) -> str:
    """Extract text from an image using pytesseract"""
    image = Image.open(image_path)
    text = pytesseract.image_to_string(image)
    return text

# Sample usage
sample_image_path = "path_to_sample_research_paper_image.png"
extracted_text = extract_text_from_image(sample_image_path)
print("Extracted text:", extracted_text[:200])

# 2. Vision-Language Model for Document Understanding

processor = ViltProcessor.from_pretrained("dandelin/vilt-b32-finetuned-vqa")
model = ViltForQuestionAnswering.from_pretrained("dandelin/vilt-b32-finetuned-vqa")

def vl_model_query(image_path: str, question: str) -> str:
    """Query the vision-language model"""
    image = Image.open(image_path)
    encoding = processor(image, question, return_tensors="pt")
    outputs = model(**encoding)
    logits = outputs.logits
    idx = logits.argmax(-1).item()
    return model.config.id2label[idx]

# Sample usage
question = "What is the main topic of this research paper?"
answer = vl_model_query(sample_image_path, question)
print(f"Question: {question}")
print(f"Answer: {answer}")

# 3. Information Synthesis

def synthesize_information(extracted_text: str, vl_model_output: str) -> str:
    """
    Synthesize information from extracted text and VL model output.
    In a real implementation, this would involve more sophisticated NLP techniques.
    """
    return f"Synthesized information: The paper discusses {vl_model_output}. Key extracted text: {extracted_text[:100]}..."

# Sample usage
synthesized_info = synthesize_information(extracted_text, answer)
print(synthesized_info)

# Full pipeline example
def research_paper_extraction_pipeline(image_path: str, question: str):
    extracted_text = extract_text_from_image(image_path)
    vl_model_output = vl_model_query(image_path, question)
    synthesized_info = synthesize_information(extracted_text, vl_model_output)
    return synthesized_info

# Run the pipeline
result = research_paper_extraction_pipeline(sample_image_path, "What is the main contribution of this paper?")
print("Final synthesized information:", result)