requirements.txt
# Core libraries
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2

# Web scraping and parsing
requests==2.25.1
beautifulsoup4==4.9.3

# Graph processing
networkx==2.6.2

# Image processing
Pillow==8.3.1
opencv-python==4.5.3.56

# OCR
pytesseract==0.3.8

# Machine Learning and NLP
torch==1.9.0
transformers==4.9.2
scikit-learn==0.24.2

# Vision-Language Models
torchvision==0.10.0

# Jupyter
jupyter==1.0.0
notebook==6.4.0